app/requirements.txt
celery==5.3.1
Django==4.2.3
django-celery-beat
django_rest_passwordreset==1.3
django-cors-headers
pandas
psycopg2-binary
djangorestframework
openpyxl
python-decouple==3.3
requests==2.23.0
redis==3.5.3
gunicorn==20.0.4
